requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
pymongo
python-dotenv
